Book api and UI updates
Created a new book API in attempt to combine the results and ordering of Open Library requests with the content of the books in the Google Books requests (cover image, description, publisher, etc.) As of now results not ideal.

Small UI updates like ordering reading sessions chronologically and changing spline file.